feat(cli): add caffeinate command
