+++ .changeset/caffeinate-cli.md
@@ -1,0 +1,6 @@
+---
+"@kilocode/cli": minor
+"kilo-code": minor
+---
+
+Add the `/caffeinate` command and notifications that explain when Kilo keeps the computer awake while agents work.
