+++ packages/kilo-vscode/src/kilo-provider/early-message.ts
@@ -27,12 +27,17 @@ type Ctx = {
   board: (message: Record<string, unknown>) => Promise<boolean>
   cancelBackgroundJob: (jobID: string, sessionID: string, requestID: string) => Promise<void>
   promoteBackgroundJob: (jobID: string, sessionID: string) => Promise<void>
+  caffeination: () => void
 }
 
 async function routeBackgroundMessage(
   message: { type: string; sessionID?: unknown; jobID?: unknown; requestID?: unknown },
   ctx: Ctx,
 ): Promise<boolean | undefined> {
+  if (message.type === "toggleCaffeination") {
+    ctx.caffeination()
+    return true
+  }
   if (message.type === "requestSessionBoard" || message.type === "resetSessionBoard") return ctx.board(message)
   if (message.type === "requestBackgroundJobs") {
     if (typeof message.sessionID === "string" && typeof message.requestID === "string") {
