+++ packages/opencode/src/kilocode/cli/cmd/tui/app.tsx
@@ -27,6 +27,7 @@ import { useIndexingWarnings } from "@/kilocode/cli/cmd/tui/indexing-warning"
 import { KiloTerminalTitle } from "./terminal-title"
 import type { KiloTitleIcon } from "./title-icon"
 import { Session as SessionApi } from "@/session/session"
+import { useCaffeination } from "./caffeination"
 
 // Re-export so upstream can render the route without importing directly
 export { KiloClawView } from "@/kilocode/claw/view"
@@ -270,6 +271,7 @@ export function init() {
 
   // Register Kilo Gateway commands (profile, teams, kiloclaw, remote, etc.)
   registerKiloCommands(useSDK)
+  useCaffeination()
 
   // Register auto-approve toggle
   useBindings(() => ({
